% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heavyR-package.R
\docType{package}
\name{heavyR-package}
\alias{heavyR}
\alias{heavyR-package}
\title{heavyR: heavyR}
\description{
Cours Université d'été.
}
\author{
\strong{Maintainer}: Clément Nerestan \email{clement.nerestan@u-bordeaux.fr} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{Cours}
\keyword{heavyR}
\keyword{internal}
