% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvndpf.R
\name{mvnpdf}
\alias{mvnpdf}
\title{Multivariate functions}
\usage{
mvnpdf(x, mean = rep(0, nrow(x)), varcovM = diag(nrow(x)), Log = TRUE)
}
\arguments{
\item{x}{A matrix}

\item{mean}{A mena value of matrix}

\item{varcovM}{variance-covariance matrix}

\item{Log}{Calculate log value, default is TRUE,}
}
\value{
A list containing the
}
\description{
it calculate the multivariate normal distribution functions
}
