% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnpdf.R
\name{mvnpdf}
\alias{mvnpdf}
\title{mvnpdf}
\usage{
mvnpdf(x, mean = rep(0, nrow(x)), varcovM = diag(nrow(x)), Log = TRUE)
}
\arguments{
\item{x}{Matrix of size NxP}

\item{mean}{a vector of means}

\item{varcovM}{a variance-covariance matrix}

\item{Log}{a logical parameter, with default value to \code{TRUE}}
}
\value{
\code{mat} matrix x

\code{vect}  a vector of length n of the multivariate normal distribution density values at those points.
}
\description{
Function compute all the values of the density evaluated at the n points in a single call of the function
}
\examples{
M1<-matrix(runif(36),nrow=6)
res <- heavyR::mvnpdf(M1)
res$mat
res$vect
}
